12-02-25--01

+++ TransitPort-Back/database/migrations/2025_01_21_102818_tiene.php
@@ -12,19 +12,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('tiene', function (Blueprint $table) {
-            $table->unsignedBigInteger('id_buque');
-            $table->unsignedBigInteger('id_contenedor');
-            
+            $table->unsignedInteger('id_buque');
+            $table->unsignedInteger('id_contenedor');
+
             // Definir clave primaria compuesta
             $table->primary(['id_buque', 'id_contenedor']);
-            
+
             // Definir claves foráneas
             $table->foreign('id_buque')->references('id')->on('buque')->onDelete('cascade');
             $table->foreign('id_contenedor')->references('id')->on('contenedor')->onDelete('cascade');
-            
+
             $table->timestamps();
         });
-        
+
     }
 
     /**
